requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
aiosqlite==0.19.0
httpx==0.27.0
google-generativeai==0.4.0
typer==0.9.0
rich==13.7.1
gradio==4.19.2
pytest==8.0.2
pytest-asyncio==0.23.5
websockets>=10.0,<12.0
sse-starlette==2.0.0
structlog==24.1.0
python-dotenv==1.0.1
